+++ src/binding.rs
@@ -7,5 +7,11 @@ pub struct Binding {
     #[serde(default)]
     pub comment: String,
     pub event: ClickEvent,
+
+    #[serde(skip_serializing)]
+    #[serde(default)]
     pub cmd: Vec<String>,
+
+    #[serde(default)]
+    pub cmd_str: String,
 }